Настраиваю отправку формы через ajax

+++ models/BookModel.php
@@ -14,61 +14,60 @@ class BookModel
     {
         $this->db = (new MainModel())->connect();
     }
-    public function findById($id)
-    {
-        $sth = $this->db->prepare("SELECT * FROM `book` WHERE `id` = ?");
-        $sth->execute(array($id));
-        $value = $sth->fetch(PDO::FETCH_ASSOC);
-        return $value;
-    }
-    public function insert(array $data)
+
+    public function insert($userId, $data)
     {
         try {
-            $sql = "INSERT INTO book(first_name, last_name, phone, email, photo) VALUES(:first_name, :last_name, :phone, :email, :photo)";
+            $sql = "INSERT INTO book(first_name, last_name, phone, email, photo, user_id) VALUES(:first_name, :last_name, :phone, :email, :photo, :user_id)";
             $stmt = $this->db->prepare($sql);
+            $stmt->bindParam(":user_id", $userId, PDO::PARAM_INT);
             $stmt = $this->prepareData($stmt, $data);
             $stmt->execute();
             return true;
         } catch (Exception $e) {
-            echo 'Code: '.$e->getCode() . '| Message: '.$e->getMessage();
+            echo 'Code: ' . $e->getCode() . '| Message: ' . $e->getMessage();
             return false;
         }
     }
 
-    public function delete($id)
+    public function update($id, $userId, $data)
     {
         try {
-            $sql = "DELETE FROM book WHERE id=:id";
+            $sql = "UPDATE book SET first_name=:first_name, last_name=:last_name, phone=:phone, email=:email, photo=:photo WHERE id=:id AND user_id=:user_id";
             $stmt = $this->db->prepare($sql);
             $stmt->bindParam(":id", $id, PDO::PARAM_INT);
-            $stmt->execute();
-            return true;
-        } catch (Exception $e) {
-            echo 'Code: '.$e->getCode() . '| Message: '.$e->getMessage();
-            return false;
-        }
-    }
-
-    public function update($id, $data)
-    {
-        try {
-            $sql = "UPDATE book SET first_name=:first_name, last_name=:last_name, phone=:phone, email=:email, photo=:photo WHERE id=:id";
-            $stmt = $this->db->prepare($sql);
-            $stmt->bindParam(":id", $id, PDO::PARAM_INT);
+            $stmt->bindParam(":user_id", $userId, PDO::PARAM_INT);
             $stmt = $this->prepareData($stmt, $data);
             $stmt->execute();
             return true;
         } catch (Exception $e) {
-            echo 'Code: '.$e->getCode() . '| Message: '.$e->getMessage();
+            echo 'Code: ' . $e->getCode() . '| Message: ' . $e->getMessage();
             return false;
         }
     }
 
-    public function findAll()
+    public function delete($id, $userId)
     {
-        $sth = $this->db->prepare("SELECT * FROM book");
-        $sth->execute();
-        $value = $sth->fetchAll(PDO::FETCH_ASSOC);
+        try {
+            $sql = "DELETE FROM book WHERE id=:id AND user_id=:user_id";
+            $stmt = $this->db->prepare($sql);
+            $stmt->bindParam(":id", $id, PDO::PARAM_INT);
+            $stmt->bindParam(":user_id", $userId, PDO::PARAM_INT);
+            $stmt->execute();
+            return true;
+        } catch (Exception $e) {
+            echo 'Code: ' . $e->getCode() . '| Message: ' . $e->getMessage();
+            return false;
+        }
+    }
+
+    public function findAll($userId)
+    {
+        $sql = "SELECT * FROM book WHERE user_id=:user_id";
+        $stmt = $this->db->prepare($sql);
+        $stmt->bindParam(":user_id", $userId, PDO::PARAM_INT);
+        $stmt->execute();
+        $value = $stmt->fetchAll(PDO::FETCH_ASSOC);
         return $value;
     }
